+++ lib/array.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <stddef.h>
+#include <stdint.h>
 
 #if defined (_WIN32)
 #define _DECL __declspec(dllexport)
@@ -10,7 +11,7 @@
 class ArrayNode {
 	void* data;
 	size_t size;
-	_DECL ArrayNode(size_t nelem, size_t obj_size);
+	_DECL ArrayNode(size_t nelem, uint64_t obj_size);
 	_DECL ~ArrayNode();
 	_DECL void pushback(char* obj, size_t obj_size);
 };
